requirements.txt
Django==2.2.2
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==19.9.0
Markdown==3.3.4
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==5.0.1
